Update Guava to 28.1-android
The name of the Android flavor can be a bit misleading;
it is actually meant for any user that requires Java 7
compatibility. As the Guava wiki states:
The Android flavor does not _require_ Android, it is merely Android compatible
and, to some extent, optimized for use on Android. Code depending on Guava
that wants to be compatible with both Android and JRE should use the Android flavor.
(https://github.com/google/guava/wiki/ReleasePolicy)